cleanup variety of compiler warnings
git-svn-id: https://svn.code.sf.net/p/jackcess/code/jackcess/trunk@247 f203690c-595d-4dc9-a70b-905162fa7fd2

+++ src/java/com/healthmarketscience/jackcess/ByteUtil.java
@@ -252,9 +252,9 @@ public final class ByteUtil {
       byte h = (byte) (b & 0xF0);
       h = (byte) (h >>> 4);
       h = (byte) (h & 0x0F);
-      rtn.append(HEX_CHARS[(int) h]);
+      rtn.append(HEX_CHARS[h]);
       h = (byte) (b & 0x0F);
-      rtn.append(HEX_CHARS[(int) h]);
+      rtn.append(HEX_CHARS[h]);
 
       if (formatted == true)
       {
